eSubida de todos los controller entity repository service modificados

+++ src/main/java/com/project/genomebankapi/repository/GenomeRepository.java
@@ -4,6 +4,16 @@ import com.project.genomebankapi.entity.Genome;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface GenomeRepository extends JpaRepository<Genome, Long> {
-    List<Genome> findBySpeciesId(Long speciesId);
+/**
+ * Repositorio JPA para la entidad Genome.
+ * Permite operaciones CRUD y consultas personalizadas sobre genomas.
+ */
+public interface GenomeRepository extends JpaRepository<Genome, Integer> {
+
+    /**
+     * Buscar todos los genomas asociados a una especie por su ID.
+     * @param speciesId ID de la especie
+     * @return lista de genomas pertenecientes a la especie
+     */
+    List<Genome> findBySpeciesId(Integer speciesId);
 }
